This revision include the initial version of documentation accompanying the newly developed and validated surface transport equations. The main contributions are a new surface_theory document that describes the theory of free surfaces and surface transport equations in enough detail (I hope) that the implementation is transparent. It also describes implementation details and the design decisions taken in the FluidInterface elements.
There is also a description of the rayleigh_instability_surfactant demos
based on the paper by Campana et al (2004), showing how the surfactant
transport equations can be included into the library. The equations should be 
rationalised and incorporated into the library rather than just in the driver
code, but that's for the future.

 


git-svn-id: svn+ssh://vummath.ma.man.ac.uk/home/svn-oomphlib/public/trunk@1084 595e1d82-adf1-4282-bac7-b6073fa2ca8e