Added demo driver for a two-layer soluble surfactant code
This required modification of hijacking functions so that the option to NOT
return a custom data object is possible and specified by a boolean flag in
the command line. The default is true so the behaviour in existing codes
is unchanged.

The other changes have been to remove deprecated MPI calls

MPI_Address -> MPI_Get_address
MPI_Type_struct -> MPI_Create_type_struct
and
MPI_Errhandler -> MPI_Comm_set_errhandler

These are the standard with MPI2 and shouldn't cause any problems, but need
to be tested on a variety of machines.



git-svn-id: svn+ssh://vummath.ma.man.ac.uk/home/svn-oomphlib/public/trunk@1331 595e1d82-adf1-4282-bac7-b6073fa2ca8e